feat/refactor([back]blog-service): Post CRUD features
1. Code issued when posting
2. Create an upload folder with the code issued when registering an image
3. Apply Query DSL
4. Service logic supplementation

+++ backend/src/main/java/com/blog/common/util/FileUploadUtils.java
@@ -13,17 +13,11 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Iterator;
+import java.util.List;
 
 public class FileUploadUtils {
-    public static String saveFile(String uploadDir, String fileName, MultipartFile multipartFile) throws IOException {
-        Path uploadPath = Paths.get(uploadDir);
-
-        if (!Files.exists(uploadPath)) {
-            Files.createDirectories(uploadPath);
-        }
-
+    public static String saveFile(Path uploadDir, String fileName, MultipartFile multipartFile) throws IOException {
         try (InputStream inputStream = multipartFile.getInputStream()) {
             // 이미지 압축 시작
             BufferedImage inputImage = ImageIO.read(inputStream);
@@ -31,7 +25,7 @@ public class FileUploadUtils {
             Iterator<ImageWriter> writers = ImageIO.getImageWritersByFormatName(StringUtils.getFilenameExtension(fileName));
             ImageWriter writer = writers.next();
 
-            Path compressedFilePath = uploadPath.resolve("compressed_" + fileName); // 압축된 파일 경로
+            Path compressedFilePath = uploadDir.resolve("compressed_" + fileName); // 압축된 파일 경로
             ImageOutputStream outputStream = ImageIO.createImageOutputStream(compressedFilePath.toFile());
             writer.setOutput(outputStream);
 
@@ -51,21 +45,23 @@ public class FileUploadUtils {
         }
     }
 
-    public static void cleanDir(String dir) {
-        Path dirPath = Paths.get(dir);
-
+    public static void cleanDir(Path postDir, List<String> imgFiles) {
         try {
-            Files.list(dirPath).forEach(file->{
-                if(!Files.isDirectory(file)) {
-                    try {
-                        Files.delete(file);
-                    }catch(IOException ex) {
-                        System.out.println("Could not delete file : " + file);
+            Files.list(postDir).forEach(file -> {
+                if (!Files.isDirectory(file)) {
+                    String fileName = file.getFileName().toString();
+                    if (!imgFiles.contains(fileName)) {
+                        try {
+                            Files.delete(file);
+                            System.out.println("Deleted: " + fileName);
+                        } catch (IOException ex) {
+                            System.out.println("Could not delete file: " + fileName);
+                        }
                     }
                 }
             });
-        }catch(IOException ex2) {
-            System.out.println("Could not list directory : " + dirPath);
+        } catch (IOException ex2) {
+            System.out.println("Could not list directory: " + postDir);
         }
     }
 }
